// src/gateways/recruiter/src/constants/enum.ts
export const enum CandidateRedisKeyCache {
  CANDIDATE_LIST = 'candidate_list',
  CANDIDATE_FEEDBACKS = 'candidate_feedbacks',
  CANDIDATE_FEEDBACK_ITEMS = 'candidate_feedback-items',
}

export const enum CandidateStatus {
  NEW = 'NEW',
  FED = 'FED',
  SCREENING = 'SCREENING',
  INTERVIEWING = 'INTERVIEWING',
  OFFERED = 'OFFERED',
  HIRED = 'HIRED',
  REJECTED = 'REJECTED',
  WITHDREW = 'WITHDREW',
  PROBATION = 'PROBATION',
  DRAFT = 'DRAFT',
  ARCHIVED = 'ARCHIVED',
}

export const enum FeedbackEnum {
  THUMB_DOWN = 'thumb_down',
}

export const enum CV {
  tags = 'RENEW',
}